+++ src/main/java/com/sevenmap/core/Runtime.java
@@ -1,7 +1,7 @@
 package com.sevenmap.core;
 
 import com.sevenmap.core.map.Map;
-import com.sevenmap.core.ui.UI;
+import com.sevenmap.core.uitestf.UI;
 import com.sevenmap.spinel.Engine;
 
 import org.apache.commons.cli.CommandLine;
