+++ src/main/java/com/sevenmap/core/uitestf/PrototypeGuiN.java
@@ -1,4 +1,4 @@
-package com.sevenmap.core.ui;
+package com.sevenmap.core.uitestf;
 
 import com.sevenmap.spinel.elements.gui.GuiNode;
