+++ src/main/java/com/sevenmap/core/uitestf/RuntimeMenuGui.java
@@ -1,6 +1,6 @@
-package com.sevenmap.core.ui;
+package com.sevenmap.core.uitestf;
 
-import com.sevenmap.core.ui.events.DbSearchEvent;
+import com.sevenmap.core.uitestf.events.DbSearchEvent;
 import com.sevenmap.spinel.Engine;
 import com.sevenmap.spinel.elements.gui.GuiNode;
