+++ src/main/java/com/sevenmap/core/uitestf/UI.java
@@ -1,8 +1,8 @@
-package com.sevenmap.core.ui;
+package com.sevenmap.core.uitestf;
 
 import com.sevenmap.core.Loadable;
-import com.sevenmap.core.ui.events.DbSearchEvent;
-import com.sevenmap.core.ui.events.FileLoadedEvent;
+import com.sevenmap.core.uitestf.events.DbSearchEvent;
+import com.sevenmap.core.uitestf.events.FileLoadedEvent;
 import com.sevenmap.spinel.Engine;
 import com.sevenmap.spinel.elements.RootNode;
 import com.sevenmap.spinel.elements.gui.GuiLayer;
