+++ src/main/java/com/sevenmap/core/uitestf/events/FileLoadedEvent.java
@@ -1,16 +1,16 @@
-package com.sevenmap.core.ui.events;
+package com.sevenmap.core.uitestf.events;
 
 import com.sevenmap.spinel.Engine;
 import com.sevenmap.spinel.scheduling.events.Event;
 
-public class DbSearchEvent extends Event {
+public class FileLoadedEvent extends Event {
 
-    public DbSearchEvent(Engine engine) {
+    public FileLoadedEvent(Engine engine) {
         super(engine);
     }
 
     @Override
     public boolean equals(Object obj) {
-        return obj instanceof DbSearchEvent;
+        return obj instanceof FileLoadedEvent;
     }
 }
